+++ CHANGELOG.md
@@ -4,6 +4,7 @@ v1.2.2 - (Unreleased)
  * Added support to set filters type with modifiers `[]` or `!` 
  * Added support to use page based pagination
  * Fixed issue checking permissions on specific objects
+ * Fixed fields from interface use same endpoints as interface definition
   
 v1.2.1 - 2018-07-16
 ----